backend/requirement.txt
fastapi
pydantic[email]
uvicorn[standard]
psycopg2-binary
python-dotenv